+++ src/components/LinksPanel/components/TimePickerElement/TimePickerElement.tsx
@@ -71,6 +71,7 @@ export const TimePickerElement: React.FC<Props> = ({ link, buttonSize, gridMode 
         link.alignContentPosition && alignClassMap[link.alignContentPosition]
       )}
       key={link.name}
+      icon={!link.showCustomIcons ? link.icon : undefined}
       size={buttonSize}
       fill="outline"
       onClick={() => {
@@ -88,6 +89,9 @@ export const TimePickerElement: React.FC<Props> = ({ link, buttonSize, gridMode 
       tooltip={link.name}
       {...testIds.buttonPicker.apply(link.name)}
     >
+      {link.showCustomIcons && link.customIconUrl && !!link.customIconUrl.length && (
+        <img src={link.customIconUrl} alt="" className={styles.customIcon} />
+      )}
       {link.name}
     </Button>
   );