+++ src/types/panel.ts
@@ -318,6 +318,20 @@ export interface LinkConfig {
   includeKioskMode: boolean;
 
   /**
+   * Use custom images
+   *
+   * @type {boolean}
+   */
+  showCustomIcons?: boolean;
+
+  /**
+   * Custom image url
+   *
+   * @type {string}
+   */
+  customIconUrl?: string;
+
+  /**
    * Align content position
    *
    * @type {AlignContentPositionType}