+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the **Business Links** panel are documented in this file.
 
 - Added links align customization and current dashboard selection ([#34](https://github.com/VolkovLabs/business-links/issues/34)).
 - Add kiosk mode for link configuration ([#17](https://github.com/VolkovLabs/business-links/issues/17)).
+- Add an option to use custom icons ([#36](https://github.com/VolkovLabs/business-links/issues/36)).
 
 ## [1.3.0] - 2025-05-28